requirements.txt
addict==2.4.0
africastalking==2.0.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.13.0
attrs==26.1.0
backoff==2.2.1
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.2
deepdiff==9.0.0
distro==1.9.0
docstring_parser==0.18.0
fastapi==0.136.0
fastuuid==0.14.0
filelock==3.29.0
frozenlist==1.8.0
fsspec==2026.3.0
googleapis-common-protos==1.74.0
greenlet==3.4.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
hubspot-api-client==12.0.0
huggingface_hub==1.11.0
idna==3.12
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.14.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langfuse==4.5.0
litellm==1.82.6
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
numpy==2.4.4
openai==2.32.0
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-http==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
orderly-set==5.5.0
packaging==26.1
pandas==3.0.2
playwright==1.58.0
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
pydantic==2.13.3
pydantic_core==2.46.3
pyee==13.0.1
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
resend==2.29.0
responses==0.26.0
rich==15.0.0
rpds-py==0.30.0
schema==0.7.8
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
tabulate==0.10.0
-e git+https://github.com/sierra-research/tau2-bench.git@d11a97072c49d093f7b5a3e4fe9da95b490d43ba#egg=tau2
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.45.0
wrapt==1.17.3
yarl==1.23.0
zipp==3.23.1
